Add U+1FAF7 (Leftwards pushing hand) and U+1FAF8 (Rightwards pushing hand)
Hello!
I've been posting progress pics of this on #8 and after all the feedback I think it's time I make a proper PR!

The style of the hand emojis are somewhat inconsistent, ranging from realistic to simplified. I tried to lean more on the latter style, but I'm not a particularly skilled artist so I'm still open to any feedback on this.